stride 관련 에러로 인한 데이터 타입 수정   - half -> uint or float 4byte 형식으로
#96

+++ SOCEngine/SOCEngine/Rendering/Shadow/ShadowCommon.h
@@ -21,8 +21,8 @@ namespace Rendering
 		public:
 			struct CommonParam
 			{
-				ushort	bias;
-				ushort	index;
+				float	bias;
+				uint	index;
 
 				CommonParam() : index(0), bias(0) {}
 				~CommonParam() {}
